fix: 쿼리 리포지토리 수정
